rename VideoChannel req and res to ChannelVideo
