+++ apiType/channelVideoType/channelVideo.go
@@ -1,10 +1,10 @@
-package videoChannelType
+package channelVideoType
 
-type VideoChannelRequest struct {
+type ChannelVideoRequest struct {
 	ChannelID int64 `json:"channel_id" validate:"required,gt=0"`
 	VideoID   int64 `json:"video_id" validate:"required,gt=0"`
 }
 
-type VideoChannelResponse struct {
+type ChannelVideoResponse struct {
 	VideoAdded bool `json:"video_added"`
 }
